refactor(domain.Benefits): separate getBenefit code to Benefits
Benefits:
    move getBenefit codes from EventPlannerController
